requirements.txt
absl-py==0.7.1
astor==0.8.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.0
Flask==1.1.1
Flask-SocketIO==4.2.0
Flask-Uploads==0.2.1
gast==0.2.2
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kafka-python==1.4.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
monotonic==1.5
numpy==1.17.0
olefile==0.46
opencv-python==4.1.0.25
pandas==0.25.0
Pillow==6.1.0
protobuf==3.9.0
pygogo==0.12.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-engineio==3.9.0
python-socketio==4.3.0
PyTurboJPEG==1.2.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
soupsieve==1.9.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
